requirements.txt
altgraph==0.14
asn1==2.2.0
asn1crypto==0.23.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
docopt==0.6.2
ecdsa==0.13
falcon==0.1.10
future==0.16.0
git-review==1.26.0
idna==2.6
macholib==1.8
pefile==2017.9.3
pipreqs==0.4.9
pycparser==2.18
PyInstaller==3.3
python-mimeparse==1.6.0
requests==2.18.4
scapy-python3==0.23
six==1.11.0
tlslite-ng==0.7.0
urllib3==1.22
xmltodict==0.11.0
yarg==0.1.9
